Added no-return-in-loop plugin
refs: https://github.com/TryGhost/Toolbox/issues/105

This plugin prevents the use of an empty return statement within a loop. This would only be valid if the loop is causing side-affects and needs to exit early, in which case, break should be used.

This will initally be enabled only for migrations in Ghost, but is generally a good rule.