Updated eslint-plugin-filenames to a fork
This package is unmaintained and we wanted to add some new rules so that we can
better enforce filenaming conventions in Ghost. The changes are backwards
compatible for the existing rules, only adding an extra option to match-regex.

The changeset from the original package at the time of writing is in:
https://github.com/allouis/eslint-plugin-filenames/commit/faf1b6b0c5c0